Divide dtos to the respective folders
