+++ Source/Server/TeamTasker.Server.Application/Dtos/Teams/CreateTeamDto.cs
@@ -1,19 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using TeamTasker.Server.Domain.Entities;
+using TeamTasker.Server.Application.Dtos.Users;
 
-namespace TeamTasker.Server.Application.Dtos
+namespace TeamTasker.Server.Application.Dtos.Teams
 {
-    public class ReadTeamDto
+    public class CreateTeamDto
     {
-        public int Id { get; set; }
         public string Name { get; set; } = string.Empty;
-        public string ProjectName { get; set; } = string.Empty;
-        public int ProjectId { get; set; }
         public ICollection<ReadEmployeeDto> Employees { get; set; } = default!;
         public string LeaderName { get; set; } = string.Empty;
         public int LeaderId { get; set; }
