+++ Source/Server/TeamTasker.Server.Application/Dtos/Users/ReadEmployeeDto.cs
@@ -3,11 +3,14 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using TeamTasker.Server.Application.Dtos.Comments;
+using TeamTasker.Server.Application.Dtos.Issues;
+using TeamTasker.Server.Application.Dtos.Teams;
 using TeamTasker.Server.Domain.Entities;
 
-namespace TeamTasker.Server.Application.Dtos
+namespace TeamTasker.Server.Application.Dtos.Users
 {
-    public class ReadLeaderDto
+    public class ReadEmployeeDto
     {
         public int Id { get; set; }
         public string FirstName { get; set; } = string.Empty;
@@ -15,7 +18,7 @@ namespace TeamTasker.Server.Application.Dtos
         public string Email { get; set; } = string.Empty;
         public string Role { get; set; } = string.Empty;
         public ICollection<ReadCommentDto> Notifications { get; set; } = default!;
-        public ICollection<ReadIssueDto> ReportedIssues { get; set; } = default!;
-        public string TeamName { get; set; } = string.Empty;
+        public ICollection<ReadIssueDto> AssignedIssues { get; set; } = default!;
+        public ICollection<ReadTeamDto> Teams { get; set; } = default!;
     }
 }
